quantity logic

+++ src/index.js
@@ -5,12 +5,10 @@ import { products } from './helpers/products';
 
 import 'basiclightbox/dist/basicLightbox.min.css';
 
-
 const search = document.querySelector('.js-search');
 const list = document.querySelector('.js-list');
 const favoriteArr = JSON.parse(localStorage.getItem(common.KEY_FAVORITE)) ?? [];
 const cartArr = JSON.parse(localStorage.getItem(common.KEY_CART)) ?? [];
-
 
 createMarkup(products, list);
 
@@ -21,13 +19,22 @@ function onClick(event) {
   if (event.target.classList.contains('js-info')) {
     const { id } = event.target.closest('.js-card').dataset;
     const product = findProduct(Number(id));
-createModal(product)
+    createModal(product);
   }
   if (event.target.classList.contains('js-cart')) {
     const { id } = event.target.closest('.js-card').dataset;
     const product = findProduct(Number(id));
-    cartArr.push(product);
-    localStorage.setItem(common.KEY_CART, JSON.stringify(cartArr));
+
+    const existingProductIndex = cartArr.findIndex(
+      item => item.id === product.id
+    );
+    if (existingProductIndex !== -1) {
+      cartArr[existingProductIndex].quantity++;
+      localStorage.setItem(common.KEY_CART, JSON.stringify(cartArr));
+    } else {
+      cartArr.push({ ...product, quantity: 1 });
+      localStorage.setItem(common.KEY_CART, JSON.stringify(cartArr));
+    }
   }
   if (event.target.classList.contains('js-favorite')) {
     const { id } = event.target.closest('.js-card').dataset;
